backend/requirements.txt
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.103.0
google-auth-httplib2==0.1.1
openai==0.28.1
python-dotenv==1.0.0
functions-framework==3.4.0
mysql-connector-python==8.1.0
cryptography==41.0.4
Flask
svix==1.13.0